libs/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "binsense-libs"
version = "0.0.1"
description = "code modules for Amazon bin challenge"
authors = [
    "Nitesh C <reddy.nitesh@gmail.com>",
    "Leo P <leo@groundswellai.com>",
    "Rathi A <rathianandk@gmail.com>"
]
license = "MIT"
requires-python = ">=3.11"
packages = [
  { include = "binsense" }
]

[tool.poetry.dependencies]
numpy = ">=1.26.4"
pandas = ">=2.2.1"
pyyaml = ">=6.0.1"
pillow = ">=10.2.0"
requests = ">=2.31.0"
opencv-python = ">=4.9.0.80"
scipy = ">=1.12.0"
torch = ">=2.2.1"
torchinfo = ">=1.8.0"
torchvision = ">=0.17.1"
tqdm = ">=4.66.2"
python-dateutil = ">=2.8.2"
s3transfer = ">=0.10.1"
botocore = ">=1.34.65"
boto3 = ">=1.34.65"
huggingface_hub = ">=0.21.4"
safetensors = ">=0.4.2"
roboflow = ">=roboflow-1.1.24"
uhashring = ">=2.3"


[tool.poetry.dev-dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
pylint = "^2.16.3"
pyright = "^1.1.296"
pytest = "^7.2.1"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.pyright]

[tool.pylint."messages control"]
disable = "all"
enable = ["empty-docstring", "missing-module-docstring", "missing-class-docstring", "missing-function-docstring"]
ignore = ["setup.py", "__init__.py"]
ignore-paths = ["tests"]
